Further update based on spec update
Move the tests to their own file so that it can be more easily tracked with the spec version
