+++ src/Couchbase.Lite.Tests.Shared/Couchbase.Lite.Tests.Shared.projitems
@@ -29,6 +29,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)ScopesCollections.ReplicationTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ReplicationTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)ScopeCollectionTest.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SQLiteOptionsTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)TLSIdentityTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PerfTest.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)PredictiveQueryTest.cs" />
