+++ src/Couchbase.Lite.Tests.Shared/DatabaseTest.cs
@@ -54,36 +54,6 @@ namespace Test
 #endif
 
         [Fact]
-        public unsafe void TestConfigurationDefaults()
-        {
-            var config = new DatabaseConfiguration();
-            config.FullSync.Should().BeFalse("because the default should be false");
-
-            Database.Delete("test", null);
-            using var db = new Database("test", config);
-            var c4db = db.c4db;
-            var nativeConfig = Native.c4db_getConfig2(c4db);
-            var hasFlag = (nativeConfig->flags & C4DatabaseFlags.DiskSyncFull) == C4DatabaseFlags.DiskSyncFull;
-            hasFlag.Should().BeFalse("because the flag should be present in LiteCore");
-        }
-
-        [Fact]
-        public unsafe void TestConfigurationFullSyncRoundTrip()
-        {
-            var config = new DatabaseConfiguration()
-            {
-                FullSync = true
-            };
-
-            Database.Delete("test", null);
-            using var db = new Database("test", config);
-            var c4db = db.c4db;
-            var nativeConfig = Native.c4db_getConfig2(c4db);
-            var hasFlag = (nativeConfig->flags & C4DatabaseFlags.DiskSyncFull) == C4DatabaseFlags.DiskSyncFull;
-            hasFlag.Should().BeTrue("because the flag should be present in LiteCore");
-        }
-
-        [Fact]
         public void TestSimpleN1QLQuery()
         {
             using (var d = new MutableDocument())
